requirements.txt
# Core dependencies
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0

# AI/ML dependencies with compatible versions
openai-whisper>=20231117
sentence-transformers>=2.2.2
scikit-learn>=1.5.0  # Compatible with numpy 2.x
numpy>=1.24.0,<2.0.0  # Pin to 1.x for compatibility
transformers>=4.53.0
torch>=2.0.0

# Audio processing
pyttsx3>=2.90
gtts>=2.4.0
soundfile>=0.12.1
pyaudio>=0.2.11
webrtcvad>=2.0.10

# Document processing
PyMuPDF>=1.23.0
PyPDF2>=3.0.0
yt-dlp>=2023.11.16

# NLP
spacy>=3.7.0
nltk>=3.8.1
textstat>=0.7.3
langdetect>=1.0.9
emoji>=2.8.0

# Data science
pandas>=2.1.0
plotly>=5.17.0
matplotlib>=3.9.0
seaborn>=0.13.2
faiss-cpu>=1.7.4

# Utilities
Pillow>=10.0.0
requests>=2.31.0
requests-cache>=1.1.0
python-dateutil>=2.8.2

# Voice-to-Voice AI Coach
livekit>=0.10.0